fix: NavigationDrawer の状態を states.ts で管理する
