+++ src/lib/states.ts
@@ -1,60 +1,17 @@
 import { atom, useRecoilValue, useSetRecoilState } from "recoil";
 
-const isDrawerOpenState = atom({
-  key: "isDrawerOpenState",
+const navigationDrawerState = atom({
+  key: "navigationDrawerState",
   default: false,
 });
 
-export const useIsDrawerOpen = (): boolean => {
-  return useRecoilValue(isDrawerOpenState);
+export const useIsNavigationDrawerOpen = (): boolean => {
+  return useRecoilValue(navigationDrawerState);
 };
 
-type IsDrawerOpenActions = {
-  closeDrawer: () => void;
-  toggleDrawer: () => void;
-};
-
-export const useIsDrawerOpenActions = (): IsDrawerOpenActions => {
-  const setIsDrawerOpen = useSetRecoilState(isDrawerOpenState);
-
-  const closeDrawer = (): void => {
-    setIsDrawerOpen(false);
+export const useToggleNavigationDrawer = () => {
+  const setIsOpen = useSetRecoilState(navigationDrawerState);
+  return () => {
+    setIsOpen((prev) => !prev);
   };
-
-  const toggleDrawer = (): void => {
-    setIsDrawerOpen((prev) => !prev);
-  };
-
-  return {
-    closeDrawer,
-    toggleDrawer,
-  };
-};
-
-const isPomodoroPlayingState = atom({
-  key: "isPomodoroPlayingState",
-  default: false,
-});
-
-export const useIsPomodoroPlaying = (): boolean => {
-  return useRecoilValue(isPomodoroPlayingState);
-};
-
-type IsPomodoroPlayingActions = {
-  startPlayingPomodoro: () => void;
-  endPlayingPomodoro: () => void;
-};
-
-export const useIsPomodoroPlayingActions = (): IsPomodoroPlayingActions => {
-  const setIsPomodoroPlaying = useSetRecoilState(isPomodoroPlayingState);
-
-  const startPlayingPomodoro = (): void => {
-    setIsPomodoroPlaying(true);
-  };
-
-  const endPlayingPomodoro = (): void => {
-    setIsPomodoroPlaying(false);
-  };
-
-  return { startPlayingPomodoro, endPlayingPomodoro };
 };
